fix(ops): carry the represented convolution column in the fused GDN epilogue
GdnConvEpilogue::store published __float2bfloat16_rn(p) and then carried the
raw accumulator p to the next column, while the window a call starts from is
read out of BF16. Restarting the Op from the state it published for column j-1
therefore did not reproduce column j: 168 mismatches over both publish forms at
T = 2, 3, 4, 8, 16, and zero on the materialized route, which round-trips the
projection through a BF16 scratch and so already satisfied the invariant.

docs/maintainer/replayssm-gdn.md makes the verify-to-fold commit an exact gather
provided the record is the same BF16 represented column the baseline writes to
history. Round the carry so it is.

Adds tests/ops/test_gdn_input_proj_conv_restart.cpp, which checks restart
equivalence directly and carries the two controls that place the fault: the
projection of a column does not depend on its position in the call, and the
materialized route reports zero before and after.

+++ src/ops/gdn_input_proj/gdn_conv.cuh
@@ -113,7 +113,9 @@ struct GdnConvEpilogue {
             publish.publish(token, batch_row, row, s1, s2, p);
             s0 = s1;
             s1 = s2;
-            s2 = p;
+            // Carry forward the column that the next column would read back, not the wider
+            // accumulator: the published window is a sequence of BF16 represented columns.
+            s2 = __bfloat162float(__float2bfloat16_rn(p));
         }
     }
 };
